+++ db/taskSeed.json
@@ -1,0 +1,50 @@
+[
+  {
+    "projectID": "Invalid ID",
+    "dev": "Stan",
+    "completed": false,
+    "subtask": false,
+    "subtaskArr": [],
+    "masterTask": null
+  },
+  {
+    "projectID": "Another ID",
+    "dev": "Mabe",
+    "completed": true,
+    "subtask": false,
+    "subtaskArr": [],
+    "masterTask": null
+  },
+  {
+    "projectID": "Look at my ID",
+    "dev": "Tad",
+    "completed": true,
+    "subtask": false,
+    "subtaskArr": [],
+    "masterTask": null
+  },
+  {
+    "projectID": "No way home",
+    "dev": "Spider-man",
+    "completed": false,
+    "subtask": false,
+    "subtaskArr": [],
+    "masterTask": null
+  },
+  {
+    "projectID": "Tonari no",
+    "dev": "Totoro",
+    "completed": false,
+    "subtask": false,
+    "subtaskArr": [],
+    "masterTask": null
+  },
+  {
+    "projectID": "Everyone loves me",
+    "dev": "Panda",
+    "completed": false,
+    "subtask": false,
+    "subtaskArr": [],
+    "masterTask": null
+  }
+]
